editing saved interest...

+++ frontend/atlas/src/components/Profile/Undergraduates/Saved/UndergraduatesSavedInterest.jsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react'
-import '../../../Undergraduates/Results/undergraduatesResults.css'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import ExpandLessIcon from '@mui/icons-material/ExpandLess';
 import GroupsIcon from '@mui/icons-material/Groups';
@@ -9,13 +8,15 @@ import UndergraduatesResultStartDuration from '../../../Undergraduates/Results/U
 import UndergraduatesResultEspaSalary from '../../../Undergraduates/Results/UndergraduatesResultEspaSalary';
 import UndergraduatesResultDescription from '../../../Undergraduates/Results/UndergraduatesResultDescription';
 import './undergraduatesSaved.css'
-import PictureAsPdfIcon from '@mui/icons-material/PictureAsPdf';
 import axios from 'axios';
+import { addMarks } from '../../../Undergraduates/Results/interests';
+import UndergraduatesSavedForm from './UndergraduatesSavedForm';
 
 const UndergraduatesSavedInterest = (props) => {
 
     const {
         interest_id,
+        internship_id,
         area,
         companyName,
         town,
@@ -32,6 +33,7 @@ const UndergraduatesSavedInterest = (props) => {
         university,
         interest_description,
         marks_name,
+        marks,
         internships,
         setInternships
     } = props;
@@ -39,27 +41,59 @@ const UndergraduatesSavedInterest = (props) => {
     const [open, setOpen] = useState(false);
     const [error, setError] = useState([]);
     const [isSubmitted, setIsSubmitted] = useState(false);
+    const [file, setFile] = useState(marks ? marks : null);
+    const [fileName, setFileName] = useState(marks_name ? marks_name : []);
+    const [selectedDescription, setSelectedDescription] = useState(interest_description ?
+        interest_description : []);
 
-    const handleDelete = async () => {
-        setIsSubmitted(true);
-        setError('Η αίτηση διαγράφηκε με επιτυχία!');
-        try {
-            await axios.delete(`http://localhost:8080/interests/id=${interest_id}`);
-            const id = localStorage.getItem('id');
-            internships.forEach((element) => {
-                axios.get(`http://localhost:8080/interests/saved/undergraduate_id=${id}`)
-                .then((response) => {
-                    const data = response.data;
-                    const find = data.find((el) => element.id === el.internship_id);
-                    if(!find) {
-                        // Remove from active internships in profile
-                        setInternships((current) => current.filter((curr) => curr.internship_id !== element.internship_id));
-                    }
-                })
-            })
+    const handleSubmit = async (e) => {
+        e.preventDefault();
+        if(isSubmitted) {
+            setError('Η αίτηση υποβλήθηκε με επιτυχία!');
+            const interest = {
+                "id": interest_id,
+                "undergraduate_id": localStorage.getItem('id'),
+                "internship_id": internship_id,
+                "description": selectedDescription.length !== 0 ? selectedDescription : null,
+                "marks_name": fileName.length !== 0 ? fileName : null,
+                "status": "await",
+                "submitted": true
+            }
+            try {
+                const response = await axios.put('http://localhost:8080/interests', interest);
+                const data = response.data;
+                console.log(data);
+            }
+            catch(error) {
+                console.error(error);
+            }
+            if(file && fileName !== marks_name) {
+                const marks = new FormData();
+                marks.append('marks', file);
+                await addMarks(interest_id, marks);
+            }
+            setInternships((current) => current.filter((curr) => curr.internship_id !== internship_id));
         }
-        catch(error) {
-            console.error(error);
+        else {
+            setError('Η αίτηση διαγράφηκε με επιτυχία!');
+            try {
+                await axios.delete(`http://localhost:8080/interests/id=${interest_id}`);
+                const id = localStorage.getItem('id');
+                internships.forEach((element) => {
+                    axios.get(`http://localhost:8080/interests/saved/undergraduate_id=${id}`)
+                    .then((response) => {
+                        const data = response.data;
+                        const find = data.find((el) => element.id === el.internship_id);
+                        if(!find) {
+                            // Remove from active internships in profile
+                            setInternships((current) => current.filter((curr) => curr.internship_id !== element.internship_id));
+                        }
+                    })
+                })
+            }
+            catch(error) {
+                console.error(error);
+            }
         }
     }
 
@@ -122,60 +156,21 @@ const UndergraduatesSavedInterest = (props) => {
                             <UndergraduatesResultDescription
                             description={description}
                             />
-                            <div className="undergraduates-saved-container">
-                                <div className="undergraduates-saved-wrapper">
-                                    <div className="undergraduates-saved-header">
-                                        <h1>
-                                            Η δήλωσή μου
-                                        </h1>
-                                    </div>
-                                    {
-                                        marks_name &&
-                                            <div className="undergraduates-saved-pdf">
-                                                <div className="undergraduates-saved-pdf-text">
-                                                    <p>
-                                                        {marks_name}
-                                                    </p>
-                                                </div>
-                                                <div className="undergraduates-saved-pdf-icon">
-                                                    <i
-                                                    onClick={() => window.open(`http://localhost:8080/interests/marks/${parseInt(interest_id)}/${marks_name}`)}
-                                                    >
-                                                        <PictureAsPdfIcon/>
-                                                    </i>
-                                                </div>
-                                            </div>
-                                    }
-                                    {
-                                        interest_description && 
-                                            <div className="undergraduates-saved-description">
-                                                <p>
-                                                    {interest_description}
-                                                </p>
-                                            </div>
-                                    }
-                                    {
-                                        !isSubmitted ?
-                                            <div className="undergraduates-saved-delete">
-                                                <button
-                                                type='button'
-                                                onClick={handleDelete}
-                                                >
-                                                    Διαγραφή Αίτησης
-                                                </button>
-                                            </div> :
-                                            <div className="undergraduates-saved-notification">
-                                                <p>
-                                                    {error}
-                                                </p>
-                                            </div>
-                                    }
-                                    
-                                </div>
-                            <div className="undergraduates-results-result-candidate-after" 
-                            style={{marginTop: '15px'}}
-                            />
-                            </div>
+                            <form
+                            style={{width: '100%'}}
+                            onSubmit={handleSubmit}
+                            >
+                                <UndergraduatesSavedForm
+                                fileName={fileName}
+                                setFileName={setFileName}
+                                setFile={setFile}
+                                selectedDescription={selectedDescription}
+                                setSelectedDescription={setSelectedDescription}
+                                setIsSubmitted={setIsSubmitted}
+                                error={error}
+                                />
+                            </form>
+                            <div className="undergraduates-results-result-candidate-after" />
                         </div>
                 }
             </div>
@@ -188,7 +183,7 @@ const UndergraduatesSavedInterest = (props) => {
                 {
                     open ? 
                         <ExpandLessIcon fontSize='large'/> :
-                        <ExpandMoreIcon fontSize='large'/>
+                            <ExpandMoreIcon fontSize='large'/>
                 }
             </button>
         </div>
